Split out files for user and sessions
+ Add NullSession and Session to session.py
+ Add User to user.py
